+++ StocksAppSolution/StocksApp.Core/Domain/Entities/User.cs
@@ -6,7 +6,7 @@
         public string? UserName { get; set; }
         public string? Email { get; set; }
         public string? PasswordHash { get; set; }
-        public decimal CashBalance { get; set; }
+        public double CashBalance { get; set; }
         public string? RefreshToken { get; set; }
         public DateTime? RefreshTokenExpiry { get; set; }
         public ICollection<BuyOrder>? BuyOrders { get; set; }
